requirements.txt
Django==2.0.9
djangorestframework==3.9.0
pytz==2018.5
